Revert "fix: refactor into separate files (#10)"
This reverts commit e363239093fbb2ac07c2409d09c008ec2c5546b7.
