--- src/utils/audio.ts
@@ -1,3 +1,0 @@
-import player from "play-sound";
-
-export const audioPlayer = player({});
